JOSS readiness: library-first pipeline, validation, and submission materials
- Add mes_core.pipeline with ONNX ensemble + fitted weights bundle
- CLI: mes score, validate, fit-weights; scripts/fit_mes_weights.py
- Eval harness comparing MES to posterior-only baselines
- Rest-block baseline heuristics; backend uses score_epochs
- paper.md, paper.bib, CITATION.cff, CONTRIBUTING, CoC, CHANGELOG
- Docs: joss-roadmap, cli, comparison, preprint outline
- Ruff-clean CI; bump version to 0.2.0

+++ mes_core/features/bandpower.py
@@ -6,7 +6,7 @@ positive values = power reduction = desynchronization = motor engagement.
 
 from __future__ import annotations
 
-from typing import Mapping
+from collections.abc import Mapping
 
 import numpy as np
 from scipy.signal import welch
@@ -30,7 +30,7 @@ def band_power(
         raise ValueError("data must have at least 2 dims (channels, time)")
     n_times = data.shape[-1]
     if nperseg is None:
-        nperseg = min(n_times, int(round(sfreq)))
+        nperseg = min(n_times, round(sfreq))
     nperseg = max(8, min(nperseg, n_times))
 
     freqs, psd = welch(
